+++ CHANGELOG.md
@@ -6,3 +6,4 @@ This changelog only contains the changes that are unreleased. For changes for in
 ## 3.4.1.2
 
 - Fix exported instances including disabled mods [#420]
+- Add in version field when exporting an instance [#419]